// src/runnersQcApp/shared/types.ts
export type TPerson = {
  hisaPersonId: string;
  hisaPersonName: string;
};

export type TFormFields = {
  horses: THorse[] | null;
  recType: string;
  conditionTreated?: string;
  drugName?: string;
  drugRoute?: string;
  drugDosage?: string;
  measure?: string;
  date?: string;
  time?: string;
  vaccine?: string;
  procedure?: string;
  surgery?: string;
  dental?: string;
  physiotherapy?: string;
  chiropractic?: string;
  notes?: string;
  limbTreated?: string;
  structure?: string;
  prescribingVet?: string;
  description?: string;
  clearedToWork?: boolean;
  clearedToRace?: boolean;
  inspectionType?: string;
  treatmentLocation?: TLocation;
  testResults?: string;
};

export type TFieldName = keyof TFormFields;

export type TFormByProtocolData = TFormFields;

export type TFunc = ReturnType<typeof useTranslations>;

export const HorseSchema = z.object({
  hisaHorseId: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  name: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  yearOfBirth: z.number().optional(),
  damName: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  ownerHisaId: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  ownerName: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  responsiblePersonHisaId: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  responsiblePersonName: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  attendingVet: z.array(z.string()).optional(),
  attendingVetName: z.array(z.string()).optional(),
  locationId: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  locationName: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  canRace: z.boolean().optional(),
  canRaceReason: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  canWork: z.boolean().optional(),
  canWorkReason: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
  lastUpdate: z
    .string()
    .optional()
    .nullable()
    .transform((val) => val ?? ""),
});

export type THorse = z.infer<typeof HorseSchema>;